Bugfix: Allow ID Strings
A user, Martin, reached out about a bug with IDs specifically with
quoting. Dotavious initially just used the string provided by caller
however this is problematic for a few different scenarios amongst others

1. Strings with spaces
2. Double-quoted strings

Details about IDs can be found at https://graphviz.org/doc/info/lang.html

This commit conditional formats IDs based on the following

1. Wrap string in quotes if there is a blank
2. Escape quotes found in string

My rationale for conditionally formatting is mostly around keeping
previous behavior intact. Quote from the link above

An ID is just a string; the lack of quote characters in the first two forms is just for simplicity.
There is no semantic difference between abc_2 and "abc_2", or between 2.34 and "2.34"

Given these IDs are semantically the same I wanted to keep the same
formatting which is unquoted.

NOTE: This does not address HTML strings as IDs which needs to be
addressed at a later date.